24/11/25/17:23 -성민이형이랑 충돌 방지용 통합 push-재영

+++ src/main/java/web/mvc/dto/GarakDTO.java
@@ -13,5 +13,5 @@ public class GarakDTO {
     private int garak_count;
     private String garak_type;
     private String garak_grade;
-    private int garak_category;
+    private long garak_category;
 }
